+++ Cesium3DTilesSelection/src/GltfContent.cpp
@@ -5,9 +5,11 @@
 #include <CesiumAsync/IAssetResponse.h>
 #include <CesiumGeometry/Axis.h>
 #include <CesiumGeometry/AxisTransforms.h>
+#include <CesiumGeospatial/BoundingRegionBuilder.h>
 #include <CesiumGltf/AccessorView.h>
 #include <CesiumGltf/AccessorWriter.h>
 #include <CesiumGltf/Model.h>
+#include <CesiumGltfReader/GltfReader.h>
 #include <CesiumUtility/Math.h>
 #include <CesiumUtility/Tracing.h>
 #include <CesiumUtility/joinToString.h>
@@ -15,15 +17,16 @@
 #include <optional>
 #include <stdexcept>
 
+using namespace CesiumAsync;
 using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
 using namespace CesiumGltf;
+using namespace CesiumGltfReader;
+using namespace CesiumUtility;
 
 namespace Cesium3DTilesSelection {
 
-using namespace CesiumAsync;
-
-/*static*/ CesiumGltf::GltfReader GltfContent::_gltfReader{};
+/*static*/ CesiumGltfReader::GltfReader GltfContent::_gltfReader{};
 
 Future<std::unique_ptr<TileContentLoadResult>>
 GltfContent::load(const TileContentLoadInput& input) {
@@ -73,7 +76,7 @@ Future<std::unique_ptr<TileContentLoadResult>> GltfContent::load(
     loadedModel.model.value().extras["Cesium3DTiles_TileUrl"] = url;
   }
 
-  return CesiumGltf::GltfReader::resolveExternalData(
+  return CesiumGltfReader::GltfReader::resolveExternalData(
              asyncSystem,
              url,
              headers,
@@ -81,7 +84,7 @@ Future<std::unique_ptr<TileContentLoadResult>> GltfContent::load(
              std::move(loadedModel),
              readOptions)
       .thenInWorkerThread(
-          [pLogger, url](CesiumGltf::ModelReaderResult&& resolvedModel) {
+          [pLogger, url](CesiumGltfReader::ModelReaderResult&& resolvedModel) {
             std::unique_ptr<TileContentLoadResult> pResult =
                 std::make_unique<TileContentLoadResult>();
 
@@ -105,49 +108,6 @@ Future<std::unique_ptr<TileContentLoadResult>> GltfContent::load(
             return pResult;
           });
 }
-
-namespace {
-void updateBoundsWithNewPosition(
-    const Cartographic& position,
-    bool haveFirst,
-    double& west,
-    double& south,
-    double& east,
-    double& north,
-    double& minimumHeight,
-    double& maximumHeight) {
-  double longitude = position.longitude;
-  double latitude = position.latitude;
-  double ellipsoidHeight = position.height;
-
-  if (haveFirst) {
-    // Check if we've crossed the anti-meridian relative to the first
-    // point and if so adjust so our min/maxes below have the expected
-    // result.
-    double difference = longitude - west;
-    if (difference > Math::ONE_PI) {
-      longitude -= Math::TWO_PI;
-    } else if (difference < -Math::ONE_PI) {
-      longitude += Math::TWO_PI;
-    }
-  } else {
-    haveFirst = true;
-  }
-
-  // The computation of longitude is very unstable at the poles,
-  // so don't let extreme latitudes affect the longitude bounding box.
-  if (glm::abs(glm::abs(latitude) - CesiumUtility::Math::PI_OVER_TWO) >
-      CesiumUtility::Math::EPSILON6) {
-    west = glm::min(west, longitude);
-    east = glm::max(east, longitude);
-  }
-  south = glm::min(south, latitude);
-  north = glm::max(north, latitude);
-  minimumHeight = glm::min(minimumHeight, ellipsoidHeight);
-  maximumHeight = glm::max(maximumHeight, ellipsoidHeight);
-}
-
-} // namespace
 
 /*static*/ std::optional<TileContentDetailsForOverlays>
 GltfContent::createRasterOverlayTextureCoordinates(
@@ -187,13 +147,11 @@ GltfContent::createRasterOverlayTextureCoordinates(
   std::vector<int> positionAccessorsToTextureCoordinateAccessor;
   positionAccessorsToTextureCoordinateAccessor.resize(gltf.accessors.size(), 0);
 
-  double west = CesiumUtility::Math::ONE_PI;
-  double south = CesiumUtility::Math::PI_OVER_TWO;
-  double east = -CesiumUtility::Math::ONE_PI;
-  double north = -CesiumUtility::Math::PI_OVER_TWO;
-  double minimumHeight = std::numeric_limits<double>::max();
-  double maximumHeight = std::numeric_limits<double>::lowest();
-  bool haveFirst = false;
+  // When computing the tile's bounds, ignore tiles that are less than 1/1000th
+  // of a tile width from the North or South pole. Longitudes cannot be trusted
+  // at such extreme latitudes.
+  BoundingRegionBuilder computedBounds;
+  computedBounds.setPoleTolerance(0.001 * bounds.computeHeight());
 
   auto createTextureCoordinatesForPrimitive =
       [&](CesiumGltf::Model& gltf,
@@ -313,15 +271,7 @@ GltfContent::createRasterOverlayTextureCoordinates(
             continue;
           }
 
-          updateBoundsWithNewPosition(
-              *cartographic,
-              haveFirst,
-              west,
-              south,
-              east,
-              north,
-              minimumHeight,
-              maximumHeight);
+          computedBounds.expandToIncludePosition(*cartographic);
 
           // Generate texture coordinates at this position for each projection
           for (size_t projectionIndex = 0; projectionIndex < projections.size();
@@ -393,10 +343,7 @@ GltfContent::createRasterOverlayTextureCoordinates(
   return TileContentDetailsForOverlays{
       std::move(projections),
       std::move(rectangles),
-      CesiumGeospatial::BoundingRegion(
-          CesiumGeospatial::GlobeRectangle(west, south, east, north),
-          minimumHeight,
-          maximumHeight)};
+      computedBounds.toRegion()};
 }
 
 /*static*/ CesiumGeospatial::BoundingRegion GltfContent::computeBoundingRegion(
@@ -409,24 +356,14 @@ GltfContent::createRasterOverlayTextureCoordinates(
   rootTransform = applyRtcCenter(gltf, rootTransform);
   rootTransform = applyGltfUpAxisTransform(gltf, rootTransform);
 
-  double west = CesiumUtility::Math::ONE_PI;
-  double south = CesiumUtility::Math::PI_OVER_TWO;
-  double east = -CesiumUtility::Math::ONE_PI;
-  double north = -CesiumUtility::Math::PI_OVER_TWO;
-  double minimumHeight = std::numeric_limits<double>::max();
-  double maximumHeight = std::numeric_limits<double>::lowest();
-  bool haveFirst = false;
+  // When computing the tile's bounds, ignore tiles that are less than 1/1000th
+  // of a tile width from the North or South pole. Longitudes cannot be trusted
+  // at such extreme latitudes.
+  BoundingRegionBuilder computedBounds;
 
   gltf.forEachPrimitiveInScene(
       -1,
-      [&rootTransform,
-       &west,
-       &south,
-       &east,
-       &north,
-       &minimumHeight,
-       &maximumHeight,
-       &haveFirst](
+      [&rootTransform, &computedBounds](
           const CesiumGltf::Model& gltf_,
           const CesiumGltf::Node& /*node*/,
           const CesiumGltf::Mesh& /*mesh*/,
@@ -466,49 +403,11 @@ GltfContent::createRasterOverlayTextureCoordinates(
             continue;
           }
 
-          double longitude = cartographic->longitude;
-          double latitude = cartographic->latitude;
-          double height = cartographic->height;
-
-          if (haveFirst) {
-            // Check if we've crossed the anti-meridian relative to the first
-            // point and if so adjust so our min/maxes below have the expected
-            // result.
-            double difference = longitude - west;
-            if (difference > Math::ONE_PI) {
-              longitude -= Math::TWO_PI;
-            } else if (difference < -Math::ONE_PI) {
-              longitude += Math::TWO_PI;
-            }
-          } else {
-            haveFirst = true;
-          }
-
-          // The computation of longitude is very unstable at the poles,
-          // so don't let extreme latitudes affect the longitude bounding box.
-          if (glm::abs(
-                  glm::abs(cartographic->latitude) -
-                  CesiumUtility::Math::PI_OVER_TWO) >
-              CesiumUtility::Math::EPSILON6) {
-            west = glm::min(west, longitude);
-            east = glm::max(east, longitude);
-          }
-          south = glm::min(south, latitude);
-          north = glm::max(north, latitude);
-          minimumHeight = glm::min(minimumHeight, height);
-          maximumHeight = glm::max(maximumHeight, height);
+          computedBounds.expandToIncludePosition(*cartographic);
         }
       });
 
-  // Put longitudes back in the -PI to PI range, which may make east < west but
-  // that's ok.
-  west = Math::negativePiToPi(west);
-  east = Math::negativePiToPi(east);
-
-  return CesiumGeospatial::BoundingRegion(
-      CesiumGeospatial::GlobeRectangle(west, south, east, north),
-      minimumHeight,
-      maximumHeight);
+  return computedBounds.toRegion();
 }
 
 /*static*/ glm::dmat4x4 GltfContent::applyRtcCenter(