+++ CesiumGltfReader/src/GltfReader.cpp
@@ -1,23 +1,17 @@
-#include "CesiumGltf/GltfReader.h"
+#include "CesiumGltfReader/GltfReader.h"
 
-#include "CesiumAsync/IAssetRequest.h"
-#include "CesiumAsync/IAssetResponse.h"
-#include "CesiumJsonReader/JsonHandler.h"
-#include "CesiumJsonReader/JsonReader.h"
-#include "CesiumUtility/Tracing.h"
-#include "CesiumUtility/Uri.h"
-#include "ExtensionKhrDracoMeshCompressionJsonHandler.h"
-#include "ExtensionMeshPrimitiveExtFeatureMetadataJsonHandler.h"
-#include "ExtensionModelExtFeatureMetadataJsonHandler.h"
-#include "ExtensionTextureBasisUJsonHandler.h"
 #include "ModelJsonHandler.h"
 #include "decodeDataUrls.h"
 #include "decodeDraco.h"
+#include "registerExtensions.h"
 
+#include <CesiumAsync/IAssetRequest.h>
+#include <CesiumAsync/IAssetResponse.h>
 #include <CesiumJsonReader/ExtensionReaderContext.h>
 #include <CesiumJsonReader/JsonHandler.h>
 #include <CesiumJsonReader/JsonReader.h>
 #include <CesiumUtility/Tracing.h>
+#include <CesiumUtility/Uri.h>
 
 #include <ktx.h>
 #include <rapidjson/reader.h>
@@ -34,11 +28,11 @@
 
 using namespace CesiumAsync;
 using namespace CesiumGltf;
+using namespace CesiumGltfReader;
 using namespace CesiumJsonReader;
 using namespace CesiumUtility;
 
 namespace {
-
 #pragma pack(push, 1)
 struct GlbHeader {
   uint32_t magic;
@@ -64,10 +58,11 @@ ModelReaderResult readJsonModel(
     const CesiumJsonReader::ExtensionReaderContext& context,
     const gsl::span<const std::byte>& data) {
 
-  CESIUM_TRACE("CesiumGltf::ModelReader::readJsonModel");
+  CESIUM_TRACE("CesiumGltfReader::ModelReader::readJsonModel");
 
   ModelJsonHandler modelHandler(context);
-  ReadJsonResult<Model> jsonResult = JsonReader::readJson(data, modelHandler);
+  CesiumJsonReader::ReadJsonResult<CesiumGltf::Model> jsonResult =
+      CesiumJsonReader::JsonReader::readJson(data, modelHandler);
 
   return ModelReaderResult{
       std::move(jsonResult.value),
@@ -75,7 +70,6 @@ ModelReaderResult readJsonModel(
       std::move(jsonResult.warnings)};
 }
 
-namespace {
 /**
  * @brief Creates a string representation for the given magic value.
  *
@@ -95,12 +89,11 @@ std::string toMagicString(uint32_t i) {
   stream << c0 << c1 << c2 << c3 << " (0x" << std::hex << i << ")";
   return stream.str();
 }
-} // namespace
 
 ModelReaderResult readBinaryModel(
     const CesiumJsonReader::ExtensionReaderContext& context,
     const gsl::span<const std::byte>& data) {
-  CESIUM_TRACE("CesiumGltf::ModelReader::readBinaryModel");
+  CESIUM_TRACE("CesiumGltfReader::ModelReader::readBinaryModel");
 
   if (data.size() < sizeof(GlbHeader) + sizeof(ChunkHeader)) {
     return {std::nullopt, {"Too short to be a valid GLB."}, {}};
@@ -190,7 +183,7 @@ ModelReaderResult readBinaryModel(
   ModelReaderResult result = readJsonModel(context, jsonChunk);
 
   if (result.model && !binaryChunk.empty()) {
-    Model& model = result.model.value();
+    CesiumGltf::Model& model = result.model.value();
 
     if (model.buffers.empty()) {
       result.errors.emplace_back(
@@ -198,7 +191,7 @@ ModelReaderResult readBinaryModel(
       return result;
     }
 
-    Buffer& buffer = model.buffers[0];
+    CesiumGltf::Buffer& buffer = model.buffers[0];
     if (buffer.uri) {
       result.errors.emplace_back("GLB has a binary chunk but the first buffer "
                                  "in the JSON chunk also has a 'uri'.");
@@ -225,7 +218,7 @@ void postprocess(
     const GltfReader& reader,
     ModelReaderResult& readModel,
     const ReadModelOptions& options) {
-  Model& model = readModel.model.value();
+  CesiumGltf::Model& model = readModel.model.value();
 
   if (options.decodeDataUrls) {
     decodeDataUrls(reader, readModel, options);
@@ -290,20 +283,7 @@ void postprocess(
 
 } // namespace
 
-GltfReader::GltfReader() : _context() {
-  this->_context.registerExtension<
-      MeshPrimitive,
-      ExtensionKhrDracoMeshCompressionJsonHandler>();
-
-  this->_context
-      .registerExtension<Model, ExtensionModelExtFeatureMetadataJsonHandler>();
-  this->_context.registerExtension<
-      MeshPrimitive,
-      ExtensionMeshPrimitiveExtFeatureMetadataJsonHandler>();
-
-  this->_context
-      .registerExtension<Texture, ExtensionTextureBasisUJsonHandler>();
-}
+GltfReader::GltfReader() : _context() { registerExtensions(this->_context); }
 
 CesiumJsonReader::ExtensionReaderContext& GltfReader::getExtensions() {
   return this->_context;
@@ -478,7 +458,7 @@ ImageReaderResult GltfReader::readImage(
   ImageReaderResult result;
 
   result.image.emplace();
-  ImageCesium& image = result.image.value();
+  CesiumGltf::ImageCesium& image = result.image.value();
 
   if (isKtx(data)) {
     // TODO: better error handling; make sure KTX-Software doesn't throw