+++ CesiumGltfReader/src/NamedObjectJsonHandler.cpp
@@ -5,15 +5,14 @@
 
 #include <string>
 
-using namespace CesiumGltf;
-
+namespace CesiumGltfReader {
 NamedObjectJsonHandler::NamedObjectJsonHandler(
     const CesiumJsonReader::ExtensionReaderContext& context) noexcept
     : CesiumJsonReader::ExtensibleObjectJsonHandler(context), _name() {}
 
 void NamedObjectJsonHandler::reset(
     IJsonHandler* pParent,
-    NamedObject* pObject) {
+    CesiumGltf::NamedObject* pObject) {
   CesiumJsonReader::ExtensibleObjectJsonHandler::reset(pParent, pObject);
 }
 
@@ -21,9 +20,10 @@ CesiumJsonReader::IJsonHandler*
 NamedObjectJsonHandler::readObjectKeyNamedObject(
     const std::string& objectType,
     const std::string_view& str,
-    NamedObject& o) {
+    CesiumGltf::NamedObject& o) {
   using namespace std::string_literals;
   if ("name"s == str)
     return property("name", this->_name, o.name);
   return this->readObjectKeyExtensibleObject(objectType, str, o);
 }
+} // namespace CesiumGltfReader