+++ CesiumJsonReader/src/StringJsonHandler.cpp
@@ -2,8 +2,7 @@
 
 #include "CesiumJsonReader/JsonHandler.h"
 
-using namespace CesiumJsonReader;
-
+namespace CesiumJsonReader {
 StringJsonHandler::StringJsonHandler() noexcept : JsonHandler() {}
 
 void StringJsonHandler::reset(IJsonHandler* pParent, std::string* pString) {
@@ -17,3 +16,4 @@ IJsonHandler* StringJsonHandler::readString(const std::string_view& str) {
   *this->_pString = str;
   return this->parent();
 }
+} // namespace CesiumJsonReader