+++ CesiumJsonWriter/include/CesiumJsonWriter/JsonObjectWriter.h
@@ -1,11 +1,15 @@
 #pragma once
 
-#include "JsonWriter.h"
+// forward declarations
+namespace CesiumJsonWriter {
+class JsonWriter;
+}
 
-#include <CesiumUtility/JsonValue.h>
+// forward declarations
+namespace CesiumUtility {
+class JsonValue;
+}
 
 namespace CesiumJsonWriter {
-void writeJsonValue(
-    const CesiumUtility::JsonValue& value,
-    CesiumJsonWriter::JsonWriter& writer);
+void writeJsonValue(const CesiumUtility::JsonValue& value, JsonWriter& writer);
 }