+++ CesiumUtility/include/CesiumUtility/ExtensibleObject.h
@@ -38,7 +38,7 @@ struct CESIUMUTILITY_API ExtensibleObject {
    * @brief Gets a generic extension with the given name as a
    * {@link CesiumUtility::JsonValue}.
    *
-   * If the extension exists but has a static type, this method will retur
+   * If the extension exists but has a static type, this method will return
    * nullptr. Use {@link getExtension} to retrieve a statically-typed extension.
    *
    * @param extensionName The name of the extension.